requirements.txt
asgiref
backports.zoneinfo
black==22.8.0
click
Django==4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.1
drf-yasg
importlib-metadata
Markdown
mypy-extensions
pathspec
Pillow
platformdirs
psycopg2-binary
python-dotenv
pytz
sqlparse
tomli
typing-extensions
whitenoise==6.2.0

